+++ src/main/java/com/diplome/game_recommendation/dtos/RecommendationDto.java
@@ -1,6 +1,5 @@
 package com.diplome.game_recommendation.dtos;
 
-import java.math.BigDecimal;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -9,6 +8,6 @@ public class RecommendationDto {
     private Long gameId;
     private String name;
     private String posterUrl;
-    private BigDecimal rating;
+    private Double rating;
     private Double recommendationScore;
 }
